refactor: split schema.ts into domain-scoped files
Closes #377
