+++ drizzle.config.ts
@@ -6,7 +6,7 @@ config({ path: "./.env", quiet: true });
 
 export default defineConfig({
   dialect: "postgresql",
-  schema: "./lib/db/schema.ts",
+  schema: "./lib/db/schema",
   out: "./drizzle",
   dbCredentials: {
     url: process.env.DATABASE_URL!,
