--- lib/db/schema.ts
@@ -1,1515 +1,0 @@
-import {
-  bigint,
-  boolean,
-  index,
-  integer,
-  pgEnum,
-  pgTable,
-  real,
-  text,
-  timestamp,
-  unique,
-  jsonb,
-  primaryKey,
-  check,
-} from "drizzle-orm/pg-core";
-import { relations, sql } from "drizzle-orm";
-import type { ConfigSnapshot } from "@/lib/types/deploy-snapshot";
-
-// ---------------------------------------------------------------------------
-// Enums
-// ---------------------------------------------------------------------------
-
-export const sourceEnum = pgEnum("source", ["git", "direct"]);
-
-export const deployTypeEnum = pgEnum("deploy_type", [
-  "compose",
-  "dockerfile",
-  "image",
-  "static",
-  "nixpacks",
-  "railpack",
-]);
-
-export const appStatusEnum = pgEnum("app_status", [
-  "active",
-  "stopped",
-  "error",
-  "deploying",
-]);
-
-export const deploymentStatusEnum = pgEnum("deployment_status", [
-  "queued",
-  "running",
-  "success",
-  "failed",
-  "cancelled",
-  "rolled_back",
-]);
-
-export const deploymentTriggerEnum = pgEnum("deployment_trigger", [
-  "manual",
-  "webhook",
-  "api",
-  "rollback",
-]);
-
-export const environmentTypeEnum = pgEnum("environment_type", [
-  "production",
-  "staging",
-  "preview",
-]);
-
-export const cloneStrategyEnum = pgEnum("clone_strategy", [
-  "clone",
-  "clone_data",
-  "empty",
-  "skip",
-]);
-
-export const groupEnvironmentTypeEnum = pgEnum("group_environment_type", [
-  "staging",
-  "preview",
-]);
-
-export const transferStatusEnum = pgEnum("transfer_status", [
-  "pending",
-  "accepted",
-  "rejected",
-  "cancelled",
-]);
-
-export const notificationChannelTypeEnum = pgEnum("notification_channel_type", [
-  "email",
-  "webhook",
-  "slack",
-]);
-
-export const meshPeerTypeEnum = pgEnum("mesh_peer_type", [
-  "persistent",
-  "dev",
-]);
-
-export const meshPeerStatusEnum = pgEnum("mesh_peer_status", [
-  "online",
-  "offline",
-  "unreachable",
-]);
-
-// ---------------------------------------------------------------------------
-// Better Auth tables (snake_case columns, matching Scope's working schema)
-// ---------------------------------------------------------------------------
-
-export const user = pgTable("user", {
-  id: text("id").primaryKey(),
-  name: text("name"),
-  email: text("email").notNull().unique(),
-  emailVerified: boolean("email_verified").default(false).notNull(),
-  image: text("image"),
-  isAppAdmin: boolean("is_app_admin").default(false),
-  twoFactorEnabled: boolean("two_factor_enabled").default(false),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-export const session = pgTable("session", {
-  id: text("id").primaryKey(),
-  expiresAt: timestamp("expires_at").notNull(),
-  token: text("token").notNull().unique(),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  ipAddress: text("ip_address"),
-  userAgent: text("user_agent"),
-  userId: text("user_id")
-    .notNull()
-    .references(() => user.id, { onDelete: "cascade" }),
-});
-
-export const account = pgTable("account", {
-  id: text("id").primaryKey(),
-  accountId: text("account_id").notNull(),
-  providerId: text("provider_id").notNull(),
-  userId: text("user_id")
-    .notNull()
-    .references(() => user.id, { onDelete: "cascade" }),
-  accessToken: text("access_token"),
-  refreshToken: text("refresh_token"),
-  idToken: text("id_token"),
-  accessTokenExpiresAt: timestamp("access_token_expires_at"),
-  refreshTokenExpiresAt: timestamp("refresh_token_expires_at"),
-  scope: text("scope"),
-  password: text("password"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-export const verification = pgTable("verification", {
-  id: text("id").primaryKey(),
-  identifier: text("identifier").notNull(),
-  value: text("value").notNull(),
-  expiresAt: timestamp("expires_at").notNull(),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-export const passkey = pgTable("passkey", {
-  id: text("id").primaryKey(),
-  name: text("name"),
-  publicKey: text("public_key").notNull(),
-  userId: text("user_id")
-    .notNull()
-    .references(() => user.id, { onDelete: "cascade" }),
-  credentialID: text("credential_id").notNull(),
-  counter: integer("counter").notNull(),
-  deviceType: text("device_type").notNull(),
-  backedUp: boolean("backed_up").notNull(),
-  transports: text("transports"),
-  createdAt: timestamp("created_at").defaultNow(),
-});
-
-export const twoFactor = pgTable("two_factor", {
-  id: text("id").primaryKey(),
-  secret: text("secret").notNull(),
-  backupCodes: text("backup_codes").notNull(),
-  userId: text("user_id")
-    .notNull()
-    .references(() => user.id, { onDelete: "cascade" }),
-});
-
-// ---------------------------------------------------------------------------
-// Organizations
-// ---------------------------------------------------------------------------
-
-export const organizations = pgTable("organization", {
-  id: text("id").primaryKey(),
-  name: text("name").notNull(),
-  slug: text("slug").notNull().unique(),
-  logo: text("logo"),
-  baseDomain: text("base_domain"),
-  sslEnabled: boolean("ssl_enabled").default(true),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-export const memberships = pgTable(
-  "membership",
-  {
-    id: text("id").primaryKey(),
-    userId: text("user_id")
-      .notNull()
-      .references(() => user.id, { onDelete: "cascade" }),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    role: text("role").notNull(),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("membership_user_id_idx").on(t.userId),
-    index("membership_org_id_idx").on(t.organizationId),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Organization Environment Variables (shared across apps)
-// ---------------------------------------------------------------------------
-
-export const orgEnvVars = pgTable(
-  "org_env_var",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    key: text("key").notNull(),
-    value: text("value").notNull(),
-    description: text("description"),
-    isSecret: boolean("is_secret").default(false),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [unique("org_env_var_org_key_uniq").on(t.organizationId, t.key)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Organization Domains (additive domain list)
-// ---------------------------------------------------------------------------
-
-export const orgDomains = pgTable(
-  "org_domain",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    domain: text("domain").notNull(),
-    isDefault: boolean("is_default").default(false),
-    enabled: boolean("enabled").default(true).notNull(),
-    verified: boolean("verified").default(false),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [unique("org_domain_uniq").on(t.organizationId, t.domain)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Deploy Keys
-// ---------------------------------------------------------------------------
-
-export const deployKeys = pgTable("deploy_key", {
-  id: text("id").primaryKey(),
-  organizationId: text("organization_id")
-    .notNull()
-    .references(() => organizations.id, { onDelete: "cascade" }),
-  name: text("name").notNull(),
-  publicKey: text("public_key").notNull(),
-  privateKey: text("private_key").notNull(), // AES-256-GCM encrypted
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: GitHub App Installations
-// ---------------------------------------------------------------------------
-
-export const githubAppInstallations = pgTable(
-  "github_app_installation",
-  {
-    id: text("id").primaryKey(),
-    userId: text("user_id")
-      .notNull()
-      .references(() => user.id, { onDelete: "cascade" }),
-    installationId: integer("installation_id").notNull(),
-    accountLogin: text("account_login").notNull(),
-    accountType: text("account_type").notNull(), // "User" or "Organization"
-    accountAvatarUrl: text("account_avatar_url"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [
-    unique("gh_install_user_uniq").on(t.userId, t.installationId),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Projects (groups of related apps)
-// ---------------------------------------------------------------------------
-
-export const projects = pgTable(
-  "project",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    displayName: text("display_name").notNull(),
-    description: text("description"),
-    color: text("color").default("#6366f1"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [unique("project_org_name_uniq").on(t.organizationId, t.name)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Apps (deployable Docker units)
-// ---------------------------------------------------------------------------
-
-export const apps = pgTable(
-  "app",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    displayName: text("display_name").notNull(),
-    description: text("description"),
-    source: sourceEnum("source").notNull().default("git"),
-    deployType: deployTypeEnum("deploy_type").notNull().default("compose"),
-    gitUrl: text("git_url"),
-    gitBranch: text("git_branch").default("main"),
-    gitKeyId: text("git_key_id").references(() => deployKeys.id, {
-      onDelete: "set null",
-    }),
-    imageName: text("image_name"),
-    composeContent: text("compose_content"),
-    composeFilePath: text("compose_file_path").default("docker-compose.yml"),
-    rootDirectory: text("root_directory"),
-    autoTraefikLabels: boolean("auto_traefik_labels").default(false),
-    containerPort: integer("container_port"),
-    autoDeploy: boolean("auto_deploy").default(false),
-    // DEPRECATED: persistentVolumes JSONB replaced by the `volumes` table.
-    // Column retained temporarily for migration; will be dropped once all data
-    // has been migrated via `scripts/migrate-volumes.ts`.
-    persistentVolumes: jsonb("persistent_volumes").$type<
-      { name: string; mountPath: string }[]
-    >(),
-    exposedPorts: jsonb("exposed_ports").$type<
-      { internal: number; external?: number; protocol?: string; description?: string }[]
-    >(),
-    restartPolicy: text("restart_policy").default("unless-stopped"),
-    connectionInfo: jsonb("connection_info").$type<
-      { label: string; value: string; copyRef?: string }[]
-    >(),
-    projectId: text("project_id").references(() => projects.id, {
-      onDelete: "set null",
-    }),
-    cloneStrategy: cloneStrategyEnum("clone_strategy").default("clone"),
-    dependsOn: jsonb("depends_on").$type<string[]>(),
-    sortOrder: integer("sort_order").default(0),
-    templateName: text("template_name"),
-    templateVersion: text("template_version"),
-    status: appStatusEnum("status").notNull().default("stopped"),
-    needsRedeploy: boolean("needs_redeploy").default(false),
-    cpuLimit: real("cpu_limit"), // CPU cores (e.g. 0.5, 1, 2)
-    memoryLimit: integer("memory_limit"), // Memory in MB (e.g. 256, 512, 1024)
-    diskWriteAlertThreshold: bigint("disk_write_alert_threshold", { mode: "number" }), // bytes/hour, null = default 1GB
-    autoRollback: boolean("auto_rollback").default(false), // Rollback on crash after deploy
-    rollbackGracePeriod: integer("rollback_grace_period").default(60), // Seconds to monitor after deploy
-    envContent: text("env_content"), // Encrypted env file blob (AES-256-GCM)
-    // Compose decomposition: child service records point to parent compose app
-    parentAppId: text("parent_app_id"),
-    composeService: text("compose_service"), // service name from compose YAML
-    containerName: text("container_name"), // computed: {projectName}-{serviceName}-1
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [
-    unique("app_org_name_uniq").on(t.organizationId, t.name),
-    index("app_org_id_idx").on(t.organizationId),
-    index("app_parent_app_id_idx").on(t.parentAppId),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Deployments
-// ---------------------------------------------------------------------------
-
-export const deployments = pgTable("deployment", {
-  id: text("id").primaryKey(),
-  appId: text("app_id")
-    .notNull()
-    .references(() => apps.id, { onDelete: "cascade" }),
-  status: deploymentStatusEnum("status").notNull().default("queued"),
-  trigger: deploymentTriggerEnum("trigger").notNull(),
-  gitSha: text("git_sha"),
-  gitMessage: text("git_message"),
-  log: text("log"),
-  durationMs: integer("duration_ms"),
-  environmentId: text("environment_id").references(() => environments.id, {
-    onDelete: "set null",
-  }),
-  groupEnvironmentId: text("group_environment_id").references(
-    () => groupEnvironments.id,
-    { onDelete: "set null" }
-  ),
-  triggeredBy: text("triggered_by").references(() => user.id, {
-    onDelete: "set null",
-  }),
-  // Snapshot fields — captured on successful deploy for rollback
-  envSnapshot: text("env_snapshot"), // Encrypted env blob at deploy time (AES-256-GCM)
-  configSnapshot: jsonb("config_snapshot").$type<ConfigSnapshot>(),
-  rollbackFromId: text("rollback_from_id"),
-  startedAt: timestamp("started_at").defaultNow().notNull(),
-  finishedAt: timestamp("finished_at"),
-},
-  (t) => [
-    index("deployment_app_id_idx").on(t.appId),
-    index("deployment_app_started_at_idx").on(t.appId, t.startedAt),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Environment Variables
-// ---------------------------------------------------------------------------
-
-export const envVars = pgTable(
-  "env_var",
-  {
-    id: text("id").primaryKey(),
-    appId: text("app_id")
-      .notNull()
-      .references(() => apps.id, { onDelete: "cascade" }),
-    key: text("key").notNull(),
-    value: text("value").notNull(), // AES-256-GCM encrypted
-    environmentId: text("environment_id").references(() => environments.id, {
-      onDelete: "cascade",
-    }),
-    isSecret: boolean("is_secret").default(true),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [unique("env_var_app_key_env_uniq").on(t.appId, t.key, t.environmentId)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Domains
-// ---------------------------------------------------------------------------
-
-export const domains = pgTable("domain", {
-  id: text("id").primaryKey(),
-  appId: text("app_id")
-    .notNull()
-    .references(() => apps.id, { onDelete: "cascade" }),
-  domain: text("domain").notNull().unique(),
-  serviceName: text("service_name"),
-  port: integer("port"),
-  middlewares: text("middlewares"),
-  certResolver: text("cert_resolver").default("le"),
-  isPrimary: boolean("is_primary").default(false),
-  sslEnabled: boolean("ssl_enabled").default(true),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-},
-  (t) => [
-    index("domain_app_id_idx").on(t.appId),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Group Environments (staging/preview environments spanning a group)
-// ---------------------------------------------------------------------------
-
-export const groupEnvironments = pgTable(
-  "group_environment",
-  {
-    id: text("id").primaryKey(),
-    projectId: text("project_id")
-      .notNull()
-      .references(() => projects.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    type: groupEnvironmentTypeEnum("type").notNull().default("staging"),
-    sourceEnvironment: text("source_environment").default("production"),
-    prNumber: integer("pr_number"),
-    prUrl: text("pr_url"),
-    createdBy: text("created_by").references(() => user.id),
-    expiresAt: timestamp("expires_at"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [unique("group_env_project_name_uniq").on(t.projectId, t.name)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Environments
-// ---------------------------------------------------------------------------
-
-export const environments = pgTable(
-  "environment",
-  {
-    id: text("id").primaryKey(),
-    appId: text("app_id")
-      .notNull()
-      .references(() => apps.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    type: environmentTypeEnum("type").notNull().default("production"),
-    domain: text("domain"),
-    gitBranch: text("git_branch"),
-    isDefault: boolean("is_default").default(false),
-    clonedFromId: text("cloned_from_id"),
-    groupEnvironmentId: text("group_environment_id").references(
-      () => groupEnvironments.id,
-      { onDelete: "cascade" }
-    ),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [unique("env_app_name_uniq").on(t.appId, t.name)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: API Tokens
-// ---------------------------------------------------------------------------
-
-export const apiTokens = pgTable(
-  "api_token",
-  {
-    id: text("id").primaryKey(),
-    userId: text("user_id")
-      .notNull()
-      .references(() => user.id, { onDelete: "cascade" }),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    tokenHash: text("token_hash").notNull(),
-    lastUsedAt: timestamp("last_used_at"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("api_token_hash_idx").on(t.tokenHash),
-    index("api_token_user_org_idx").on(t.userId, t.organizationId),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Activities (audit trail)
-// ---------------------------------------------------------------------------
-
-export const activities = pgTable(
-  "activity",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    appId: text("app_id").references(() => apps.id, {
-      onDelete: "set null",
-    }),
-    userId: text("user_id").references(() => user.id, { onDelete: "set null" }),
-    action: text("action").notNull(),
-    metadata: jsonb("metadata"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("activity_org_created_at_idx").on(t.organizationId, t.createdAt),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Tags (flat labels for filtering)
-// ---------------------------------------------------------------------------
-
-export const tags = pgTable(
-  "tag",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    color: text("color").notNull().default("#6366f1"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [unique("tag_org_name_uniq").on(t.organizationId, t.name)]
-);
-
-export const appTags = pgTable(
-  "app_tag",
-  {
-    appId: text("app_id")
-      .notNull()
-      .references(() => apps.id, { onDelete: "cascade" }),
-    tagId: text("tag_id")
-      .notNull()
-      .references(() => tags.id, { onDelete: "cascade" }),
-  },
-  (t) => [unique("app_tag_uniq").on(t.appId, t.tagId)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Backup Targets (where backups are stored)
-// ---------------------------------------------------------------------------
-
-export const backupTargetTypeEnum = pgEnum("backup_target_type", [
-  "s3",
-  "r2",
-  "b2",
-  "ssh",
-  "local",
-]);
-
-export const backupTargets = pgTable("backup_target", {
-  id: text("id").primaryKey(),
-  organizationId: text("organization_id").references(() => organizations.id, {
-    onDelete: "cascade",
-  }),
-  name: text("name").notNull(),
-  type: backupTargetTypeEnum("type").notNull(),
-  config: jsonb("config")
-    .notNull()
-    .$type<
-      | {
-          bucket: string;
-          region: string;
-          endpoint?: string;
-          accessKeyId: string;
-          secretAccessKey: string;
-          prefix?: string;
-        }
-      | {
-          host: string;
-          port?: number;
-          username: string;
-          privateKey?: string;
-          path: string;
-        }
-      | {
-          path: string;
-        }
-    >(),
-  isDefault: boolean("is_default").default(false),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Backup Jobs (scheduled backup configurations)
-// ---------------------------------------------------------------------------
-
-export const backupStatusEnum = pgEnum("backup_status", [
-  "pending",
-  "running",
-  "success",
-  "failed",
-  "pruned",
-]);
-
-export const backupJobs = pgTable("backup_job", {
-  id: text("id").primaryKey(),
-  organizationId: text("organization_id")
-    .references(() => organizations.id, { onDelete: "cascade" }),
-  targetId: text("target_id")
-    .notNull()
-    .references(() => backupTargets.id, { onDelete: "cascade" }),
-  name: text("name").notNull(),
-  schedule: text("schedule").notNull().default("0 2 * * *"),
-  enabled: boolean("enabled").default(true).notNull(),
-  // Proxmox-style retention
-  keepAll: boolean("keep_all").default(false),
-  keepLast: integer("keep_last"),
-  keepHourly: integer("keep_hourly"),
-  keepDaily: integer("keep_daily"),
-  keepWeekly: integer("keep_weekly"),
-  keepMonthly: integer("keep_monthly"),
-  keepYearly: integer("keep_yearly"),
-  // Notification settings
-  notifyOnSuccess: boolean("notify_on_success").default(false),
-  notifyOnFailure: boolean("notify_on_failure").default(true),
-  lastRunAt: timestamp("last_run_at"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// Many-to-many: which apps are included in a backup job
-export const backupJobApps = pgTable(
-  "backup_job_app",
-  {
-    backupJobId: text("backup_job_id")
-      .notNull()
-      .references(() => backupJobs.id, { onDelete: "cascade" }),
-    appId: text("app_id")
-      .notNull()
-      .references(() => apps.id, { onDelete: "cascade" }),
-  },
-  (t) => [primaryKey({ columns: [t.backupJobId, t.appId] })]
-);
-
-// Many-to-many: direct volume links for backup jobs (system volumes, etc.)
-export const backupJobVolumes = pgTable(
-  "backup_job_volume",
-  {
-    backupJobId: text("backup_job_id")
-      .notNull()
-      .references(() => backupJobs.id, { onDelete: "cascade" }),
-    volumeId: text("volume_id")
-      .notNull()
-      .references(() => volumes.id, { onDelete: "cascade" }),
-  },
-  (t) => [primaryKey({ columns: [t.backupJobId, t.volumeId] })]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Backup History (individual backup runs)
-// ---------------------------------------------------------------------------
-
-export const backups = pgTable("backup", {
-  id: text("id").primaryKey(),
-  jobId: text("job_id")
-    .notNull()
-    .references(() => backupJobs.id, { onDelete: "cascade" }),
-  appId: text("app_id")
-    .references(() => apps.id, { onDelete: "cascade" }),
-  targetId: text("target_id")
-    .notNull()
-    .references(() => backupTargets.id, { onDelete: "cascade" }),
-  status: backupStatusEnum("status").notNull().default("pending"),
-  volumeName: text("volume_name"),
-  sizeBytes: bigint("size_bytes", { mode: "number" }),
-  storagePath: text("storage_path"),
-  checksum: text("checksum"), // sha256 hash of the archive before upload
-  log: text("log"),
-  startedAt: timestamp("started_at").defaultNow().notNull(),
-  finishedAt: timestamp("finished_at"),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Volumes (first-class volume records with integrated limits)
-// ---------------------------------------------------------------------------
-
-export const volumes = pgTable(
-  "volume",
-  {
-    id: text("id").primaryKey(),
-    appId: text("app_id")
-      .references(() => apps.id, { onDelete: "cascade" }),
-    organizationId: text("organization_id")
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(), // e.g. "data", "uploads"
-    mountPath: text("mount_path").notNull(), // e.g. "/var/lib/postgresql/data"
-    persistent: boolean("persistent").default(true).notNull(), // survives deploys
-    shared: boolean("shared").default(false).notNull(), // can be mounted by other apps in project
-    description: text("description"),
-    maxSizeBytes: bigint("max_size_bytes", { mode: "number" }), // nullable = no limit
-    warnAtPercent: integer("warn_at_percent").default(80),
-    ignorePatterns: jsonb("ignore_patterns").$type<string[]>(), // glob patterns to ignore in diff (e.g. "uploads/**")
-    driftCount: integer("drift_count").default(0), // unignored file drift after last deploy
-    // Backup strategy: "tar" (default) for file volumes, "dump" for databases
-    backupStrategy: text("backup_strategy").default("tar").notNull(),
-    // For "dump" strategy: { dumpCmd, restoreCmd } — shell commands run via docker exec
-    backupMeta: jsonb("backup_meta").$type<{ dumpCmd: string; restoreCmd: string }>(),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [
-    unique("volume_app_name_uniq").on(t.appId, t.name),
-    unique("volume_app_mount_uniq").on(t.appId, t.mountPath),
-    index("volume_app_id_idx").on(t.appId),
-    index("volume_org_id_idx").on(t.organizationId),
-    check("volume_dump_requires_meta", sql`backup_strategy != 'dump' OR backup_meta IS NOT NULL`),
-  ]
-);
-
-// DEPRECATED: kept for migration only — will be dropped after migrate-volumes.ts runs
-export const volumeLimits = pgTable("volume_limit", {
-  id: text("id").primaryKey(),
-  appId: text("app_id")
-    .notNull()
-    .references(() => apps.id, { onDelete: "cascade" })
-    .unique(),
-  maxSizeBytes: bigint("max_size_bytes", { mode: "number" }).notNull(),
-  warnAtPercent: integer("warn_at_percent").default(80),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Domain Checks (health monitoring history)
-// ---------------------------------------------------------------------------
-
-export const domainChecks = pgTable(
-  "domain_check",
-  {
-    id: text("id").primaryKey(),
-    domainId: text("domain_id")
-      .notNull()
-      .references(() => domains.id, { onDelete: "cascade" }),
-    reachable: boolean("reachable").notNull(),
-    statusCode: integer("status_code"),
-    responseTimeMs: integer("response_time_ms"),
-    error: text("error"),
-    checkedAt: timestamp("checked_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("domain_check_domain_checked_at_idx").on(t.domainId, t.checkedAt),
-  ]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Templates
-// ---------------------------------------------------------------------------
-
-export const templateCategoryEnum = pgEnum("template_category", [
-  "database",
-  "cache",
-  "monitoring",
-  "web",
-  "tool",
-  "custom",
-]);
-
-export const templates = pgTable("template", {
-  id: text("id").primaryKey(),
-  name: text("name").notNull().unique(),
-  displayName: text("display_name").notNull(),
-  description: text("description"),
-  icon: text("icon"),
-  category: templateCategoryEnum("category").notNull().default("custom"),
-  source: sourceEnum("source").notNull().default("direct"),
-  deployType: deployTypeEnum("deploy_type").notNull().default("image"),
-  imageName: text("image_name"),
-  gitUrl: text("git_url"),
-  gitBranch: text("git_branch"),
-  composeContent: text("compose_content"),
-  rootDirectory: text("root_directory"),
-  defaultPort: integer("default_port"),
-  defaultEnvVars: jsonb("default_env_vars").$type<
-    { key: string; description: string; required: boolean; defaultValue?: string }[]
-  >(),
-  defaultVolumes: jsonb("default_volumes").$type<
-    { name: string; mountPath: string; description: string }[]
-  >(),
-  defaultConnectionInfo: jsonb("default_connection_info").$type<
-    { label: string; value: string; copyRef?: string }[]
-  >(),
-  isBuiltIn: boolean("is_built_in").default(false).notNull(),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Cron Jobs (scheduled tasks)
-// ---------------------------------------------------------------------------
-
-export const cronJobTypeEnum = pgEnum("cron_job_type", [
-  "command",
-  "url",
-]);
-
-export const cronJobStatusEnum = pgEnum("cron_job_status", [
-  "success",
-  "failed",
-  "running",
-]);
-
-export const cronJobs = pgTable("cron_job", {
-  id: text("id").primaryKey(),
-  appId: text("app_id")
-    .notNull()
-    .references(() => apps.id, { onDelete: "cascade" }),
-  name: text("name").notNull(),
-  type: cronJobTypeEnum("type").notNull().default("command"),
-  schedule: text("schedule").notNull(), // cron expression
-  command: text("command").notNull(), // shell command or URL depending on type
-  enabled: boolean("enabled").default(true).notNull(),
-  lastRunAt: timestamp("last_run_at"),
-  lastStatus: cronJobStatusEnum("last_status"),
-  lastLog: text("last_log"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Cron Job Runs (execution history)
-// ---------------------------------------------------------------------------
-
-export const cronJobRunStatusEnum = pgEnum("cron_job_run_status", [
-  "success",
-  "failed",
-]);
-
-export const cronJobRuns = pgTable(
-  "cron_job_run",
-  {
-    id: text("id").primaryKey(),
-    cronJobId: text("cron_job_id")
-      .notNull()
-      .references(() => cronJobs.id, { onDelete: "cascade" }),
-    status: cronJobRunStatusEnum("status").notNull(),
-    startedAt: timestamp("started_at").notNull(),
-    completedAt: timestamp("completed_at"),
-    output: text("output"),
-    error: text("error"),
-  },
-  (t) => [index("cron_job_run_job_id_idx").on(t.cronJobId)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: App Transfers (move apps between organizations)
-// ---------------------------------------------------------------------------
-
-export const appTransfers = pgTable("app_transfer", {
-  id: text("id").primaryKey(),
-  appId: text("app_id")
-    .notNull()
-    .references(() => apps.id, { onDelete: "cascade" }),
-  sourceOrgId: text("source_org_id")
-    .notNull()
-    .references(() => organizations.id, { onDelete: "cascade" }),
-  destinationOrgId: text("destination_org_id")
-    .notNull()
-    .references(() => organizations.id, { onDelete: "cascade" }),
-  status: transferStatusEnum("status").notNull().default("pending"),
-  initiatedBy: text("initiated_by")
-    .references(() => user.id, { onDelete: "set null" }),
-  respondedBy: text("responded_by")
-    .references(() => user.id, { onDelete: "set null" }),
-  frozenRefs: jsonb("frozen_refs").$type<
-    { key: string; originalRef: string; frozenValue: string }[]
-  >(),
-  note: text("note"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  respondedAt: timestamp("responded_at"),
-});
-
-// ---------------------------------------------------------------------------
-// Vardo: Notification Channels
-// ---------------------------------------------------------------------------
-
-export const notificationChannels = pgTable(
-  "notification_channel",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id").notNull().references(() => organizations.id, { onDelete: "cascade" }),
-    name: text("name").notNull(),
-    type: notificationChannelTypeEnum("type").notNull(),
-    config: jsonb("config").notNull().$type<{ recipients: string[] } | { url: string; secret?: string } | { webhookUrl: string }>(),
-    enabled: boolean("enabled").default(true).notNull(),
-    subscribedEvents: text("subscribed_events").array().default([]).notNull(),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [index("notification_channel_org_idx").on(t.organizationId)]
-);
-
-// ---------------------------------------------------------------------------
-// Vardo: Weekly Digest Settings (per-org)
-// ---------------------------------------------------------------------------
-
-export const digestSettings = pgTable("digest_setting", {
-  id: text("id").primaryKey(),
-  organizationId: text("organization_id")
-    .notNull()
-    .unique()
-    .references(() => organizations.id, { onDelete: "cascade" }),
-  enabled: boolean("enabled").default(true).notNull(),
-  // 0 = Sunday … 6 = Saturday
-  dayOfWeek: integer("day_of_week").default(1).notNull(),
-  // 0-23 UTC
-  hourOfDay: integer("hour_of_day").default(8).notNull(),
-  lastSentAt: timestamp("last_sent_at"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Relations
-// ---------------------------------------------------------------------------
-
-export const userRelations = relations(user, ({ many }) => ({
-  sessions: many(session),
-  accounts: many(account),
-  memberships: many(memberships),
-  deployments: many(deployments),
-  apiTokens: many(apiTokens),
-  activities: many(activities),
-  githubAppInstallations: many(githubAppInstallations),
-  initiatedTransfers: many(appTransfers, { relationName: "initiatedByUser" }),
-  respondedTransfers: many(appTransfers, { relationName: "respondedByUser" }),
-}));
-
-export const organizationsRelations = relations(organizations, ({ many, one }) => ({
-  memberships: many(memberships),
-  projects: many(projects),
-  apps: many(apps),
-  deployKeys: many(deployKeys),
-  apiTokens: many(apiTokens),
-  activities: many(activities),
-  tags: many(tags),
-  backupTargets: many(backupTargets),
-  backupJobs: many(backupJobs),
-  orgEnvVars: many(orgEnvVars),
-  orgDomains: many(orgDomains),
-  outgoingTransfers: many(appTransfers, { relationName: "sourceOrg" }),
-  incomingTransfers: many(appTransfers, { relationName: "destinationOrg" }),
-  notificationChannels: many(notificationChannels),
-  invitations: many(invitations),
-  digestSetting: one(digestSettings, {
-    fields: [organizations.id],
-    references: [digestSettings.organizationId],
-  }),
-}));
-
-export const membershipsRelations = relations(memberships, ({ one }) => ({
-  user: one(user, {
-    fields: [memberships.userId],
-    references: [user.id],
-  }),
-  organization: one(organizations, {
-    fields: [memberships.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const projectsRelations = relations(projects, ({ one, many }) => ({
-  organization: one(organizations, {
-    fields: [projects.organizationId],
-    references: [organizations.id],
-  }),
-  apps: many(apps),
-  groupEnvironments: many(groupEnvironments),
-  instances: many(projectInstances),
-}));
-
-export const appsRelations = relations(apps, ({ one, many }) => ({
-  organization: one(organizations, {
-    fields: [apps.organizationId],
-    references: [organizations.id],
-  }),
-  deployKey: one(deployKeys, {
-    fields: [apps.gitKeyId],
-    references: [deployKeys.id],
-  }),
-  project: one(projects, {
-    fields: [apps.projectId],
-    references: [projects.id],
-  }),
-  parentApp: one(apps, {
-    fields: [apps.parentAppId],
-    references: [apps.id],
-    relationName: "parentChild",
-  }),
-  childApps: many(apps, { relationName: "parentChild" }),
-  deployments: many(deployments),
-  envVars: many(envVars),
-  domains: many(domains),
-  environments: many(environments),
-  activities: many(activities),
-  appTags: many(appTags),
-  backupJobApps: many(backupJobApps),
-  backups: many(backups),
-  volumes: many(volumes),
-  volumeLimit: many(volumeLimits),
-  cronJobs: many(cronJobs),
-  transfers: many(appTransfers),
-}));
-
-export const deploymentsRelations = relations(deployments, ({ one }) => ({
-  app: one(apps, {
-    fields: [deployments.appId],
-    references: [apps.id],
-  }),
-  environment: one(environments, {
-    fields: [deployments.environmentId],
-    references: [environments.id],
-  }),
-  groupEnvironment: one(groupEnvironments, {
-    fields: [deployments.groupEnvironmentId],
-    references: [groupEnvironments.id],
-  }),
-  triggeredByUser: one(user, {
-    fields: [deployments.triggeredBy],
-    references: [user.id],
-  }),
-}));
-
-export const envVarsRelations = relations(envVars, ({ one }) => ({
-  app: one(apps, {
-    fields: [envVars.appId],
-    references: [apps.id],
-  }),
-  environment: one(environments, {
-    fields: [envVars.environmentId],
-    references: [environments.id],
-  }),
-}));
-
-export const domainsRelations = relations(domains, ({ one, many }) => ({
-  app: one(apps, {
-    fields: [domains.appId],
-    references: [apps.id],
-  }),
-  domainChecks: many(domainChecks),
-}));
-
-export const domainChecksRelations = relations(domainChecks, ({ one }) => ({
-  domain: one(domains, {
-    fields: [domainChecks.domainId],
-    references: [domains.id],
-  }),
-}));
-
-export const environmentsRelations = relations(
-  environments,
-  ({ one, many }) => ({
-    app: one(apps, {
-      fields: [environments.appId],
-      references: [apps.id],
-    }),
-    groupEnvironment: one(groupEnvironments, {
-      fields: [environments.groupEnvironmentId],
-      references: [groupEnvironments.id],
-    }),
-    envVars: many(envVars),
-    deployments: many(deployments),
-  })
-);
-
-export const deployKeysRelations = relations(deployKeys, ({ one }) => ({
-  organization: one(organizations, {
-    fields: [deployKeys.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const githubAppInstallationsRelations = relations(
-  githubAppInstallations,
-  ({ one }) => ({
-    user: one(user, {
-      fields: [githubAppInstallations.userId],
-      references: [user.id],
-    }),
-  })
-);
-
-export const apiTokensRelations = relations(apiTokens, ({ one }) => ({
-  user: one(user, {
-    fields: [apiTokens.userId],
-    references: [user.id],
-  }),
-  organization: one(organizations, {
-    fields: [apiTokens.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const activitiesRelations = relations(activities, ({ one }) => ({
-  organization: one(organizations, {
-    fields: [activities.organizationId],
-    references: [organizations.id],
-  }),
-  app: one(apps, {
-    fields: [activities.appId],
-    references: [apps.id],
-  }),
-  user: one(user, {
-    fields: [activities.userId],
-    references: [user.id],
-  }),
-}));
-
-export const groupEnvironmentsRelations = relations(
-  groupEnvironments,
-  ({ one, many }) => ({
-    project: one(projects, {
-      fields: [groupEnvironments.projectId],
-      references: [projects.id],
-    }),
-    createdByUser: one(user, {
-      fields: [groupEnvironments.createdBy],
-      references: [user.id],
-    }),
-    environments: many(environments),
-    deployments: many(deployments),
-  })
-);
-
-export const tagsRelations = relations(tags, ({ one, many }) => ({
-  organization: one(organizations, {
-    fields: [tags.organizationId],
-    references: [organizations.id],
-  }),
-  appTags: many(appTags),
-}));
-
-export const appTagsRelations = relations(appTags, ({ one }) => ({
-  app: one(apps, {
-    fields: [appTags.appId],
-    references: [apps.id],
-  }),
-  tag: one(tags, {
-    fields: [appTags.tagId],
-    references: [tags.id],
-  }),
-}));
-
-export const backupTargetsRelations = relations(
-  backupTargets,
-  ({ one, many }) => ({
-    organization: one(organizations, {
-      fields: [backupTargets.organizationId],
-      references: [organizations.id],
-    }),
-    backupJobs: many(backupJobs),
-    backups: many(backups),
-  })
-);
-
-export const backupJobsRelations = relations(backupJobs, ({ one, many }) => ({
-  organization: one(organizations, {
-    fields: [backupJobs.organizationId],
-    references: [organizations.id],
-  }),
-  target: one(backupTargets, {
-    fields: [backupJobs.targetId],
-    references: [backupTargets.id],
-  }),
-  backupJobApps: many(backupJobApps),
-  backupJobVolumes: many(backupJobVolumes),
-  backups: many(backups),
-}));
-
-export const backupJobAppsRelations = relations(
-  backupJobApps,
-  ({ one }) => ({
-    backupJob: one(backupJobs, {
-      fields: [backupJobApps.backupJobId],
-      references: [backupJobs.id],
-    }),
-    app: one(apps, {
-      fields: [backupJobApps.appId],
-      references: [apps.id],
-    }),
-  })
-);
-
-export const backupJobVolumesRelations = relations(
-  backupJobVolumes,
-  ({ one }) => ({
-    backupJob: one(backupJobs, {
-      fields: [backupJobVolumes.backupJobId],
-      references: [backupJobs.id],
-    }),
-    volume: one(volumes, {
-      fields: [backupJobVolumes.volumeId],
-      references: [volumes.id],
-    }),
-  })
-);
-
-export const backupsRelations = relations(backups, ({ one }) => ({
-  job: one(backupJobs, {
-    fields: [backups.jobId],
-    references: [backupJobs.id],
-  }),
-  app: one(apps, {
-    fields: [backups.appId],
-    references: [apps.id],
-  }),
-  target: one(backupTargets, {
-    fields: [backups.targetId],
-    references: [backupTargets.id],
-  }),
-}));
-
-export const volumesRelations = relations(volumes, ({ one }) => ({
-  app: one(apps, {
-    fields: [volumes.appId],
-    references: [apps.id],
-  }),
-  organization: one(organizations, {
-    fields: [volumes.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const volumeLimitsRelations = relations(volumeLimits, ({ one }) => ({
-  app: one(apps, {
-    fields: [volumeLimits.appId],
-    references: [apps.id],
-  }),
-}));
-
-export const cronJobsRelations = relations(cronJobs, ({ one, many }) => ({
-  app: one(apps, {
-    fields: [cronJobs.appId],
-    references: [apps.id],
-  }),
-  runs: many(cronJobRuns),
-}));
-
-export const cronJobRunsRelations = relations(cronJobRuns, ({ one }) => ({
-  cronJob: one(cronJobs, {
-    fields: [cronJobRuns.cronJobId],
-    references: [cronJobs.id],
-  }),
-}));
-
-export const orgEnvVarsRelations = relations(orgEnvVars, ({ one }) => ({
-  organization: one(organizations, {
-    fields: [orgEnvVars.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const orgDomainsRelations = relations(orgDomains, ({ one }) => ({
-  organization: one(organizations, {
-    fields: [orgDomains.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-export const appTransfersRelations = relations(
-  appTransfers,
-  ({ one }) => ({
-    app: one(apps, {
-      fields: [appTransfers.appId],
-      references: [apps.id],
-    }),
-    sourceOrg: one(organizations, {
-      fields: [appTransfers.sourceOrgId],
-      references: [organizations.id],
-      relationName: "sourceOrg",
-    }),
-    destinationOrg: one(organizations, {
-      fields: [appTransfers.destinationOrgId],
-      references: [organizations.id],
-      relationName: "destinationOrg",
-    }),
-    initiatedByUser: one(user, {
-      fields: [appTransfers.initiatedBy],
-      references: [user.id],
-      relationName: "initiatedByUser",
-    }),
-    respondedByUser: one(user, {
-      fields: [appTransfers.respondedBy],
-      references: [user.id],
-      relationName: "respondedByUser",
-    }),
-  })
-);
-
-export const notificationChannelsRelations = relations(notificationChannels, ({ one }) => ({ organization: one(organizations, { fields: [notificationChannels.organizationId], references: [organizations.id] }) }));
-
-export const digestSettingsRelations = relations(digestSettings, ({ one }) => ({
-  organization: one(organizations, {
-    fields: [digestSettings.organizationId],
-    references: [organizations.id],
-  }),
-}));
-
-// ---------------------------------------------------------------------------
-// Vardo: Invitations
-// ---------------------------------------------------------------------------
-
-export const invitationScopeEnum = pgEnum("invitation_scope", [
-  "platform",
-  "org",
-  "project",
-]);
-
-export const invitationStatusEnum = pgEnum("invitation_status", [
-  "pending",
-  "accepted",
-  "expired",
-  "revoked",
-]);
-
-export const invitations = pgTable(
-  "invitation",
-  {
-    id: text("id").primaryKey(),
-    email: text("email").notNull(),
-    scope: invitationScopeEnum("scope").notNull(),
-    targetId: text("target_id"), // orgId for org scope, projectId for project scope, null for platform
-    role: text("role").notNull(), // "owner", "admin", "member"
-    status: invitationStatusEnum("status").notNull().default("pending"),
-    token: text("token").notNull().unique(),
-    invitedBy: text("invited_by").references(() => user.id, { onDelete: "set null" }),
-    expiresAt: timestamp("expires_at").notNull(),
-    acceptedAt: timestamp("accepted_at"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("invitation_target_scope_status_idx").on(t.targetId, t.scope, t.status),
-  ],
-);
-
-export const invitationsRelations = relations(invitations, ({ one }) => ({
-  inviter: one(user, {
-    fields: [invitations.invitedBy],
-    references: [user.id],
-  }),
-}));
-
-// ---------------------------------------------------------------------------
-// System settings (key-value store for setup wizard + global config)
-// ---------------------------------------------------------------------------
-
-// ---------------------------------------------------------------------------
-// Notification log — records every delivery attempt + result
-// ---------------------------------------------------------------------------
-
-export const notificationLogs = pgTable(
-  "notification_log",
-  {
-    id: text("id").primaryKey(),
-    organizationId: text("organization_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    channelId: text("channel_id").references(() => notificationChannels.id, {
-      onDelete: "set null",
-    }),
-    channelName: text("channel_name").notNull(),
-    channelType: text("channel_type").notNull(), // email, webhook, push
-    eventType: text("event_type").notNull(), // deploy.success, backup.failed, etc.
-    eventTitle: text("event_title").notNull(),
-    status: text("status").notNull(), // success, failed
-    error: text("error"), // error message if failed
-    attempt: integer("attempt").notNull().default(1),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (t) => [
-    index("notification_log_org_idx").on(t.organizationId),
-    index("notification_log_created_idx").on(t.createdAt),
-  ]
-);
-
-export const systemSettings = pgTable("system_settings", {
-  key: text("key").primaryKey(),
-  value: text("value").notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-// ---------------------------------------------------------------------------
-// Instance Mesh — peer registry (system-level, not org-scoped)
-// ---------------------------------------------------------------------------
-
-export const meshPeers = pgTable("mesh_peer", {
-  id: text("id").primaryKey(),
-  instanceId: text("instance_id").notNull().unique(),
-  name: text("name").notNull(),
-  type: meshPeerTypeEnum("type").notNull().default("persistent"),
-  status: meshPeerStatusEnum("status").notNull().default("offline"),
-  endpoint: text("endpoint"), // host:port for WireGuard (null for dev behind NAT)
-  publicKey: text("public_key").notNull().unique(),
-  allowedIps: text("allowed_ips").notNull(), // WireGuard AllowedIPs (CIDR)
-  internalIp: text("internal_ip").notNull().unique(), // WireGuard tunnel address (e.g. 10.99.0.1)
-  apiUrl: text("api_url"), // reachable URL for mesh API calls over tunnel
-  tokenHash: text("token_hash").unique(), // SHA-256 hash of the token we gave this peer (inbound auth)
-  outboundToken: text("outbound_token"), // token the peer gave us for calling their API (outbound auth)
-  lastSeenAt: timestamp("last_seen_at"),
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at").defaultNow().notNull(),
-});
-
-export const meshPeersRelations = relations(meshPeers, ({ many }) => ({
-  projectInstances: many(projectInstances),
-}));
-
-// ---------------------------------------------------------------------------
-// Instance Mesh — project-to-instance-environment mapping
-// ---------------------------------------------------------------------------
-
-export const projectInstances = pgTable(
-  "project_instance",
-  {
-    id: text("id").primaryKey(),
-    projectId: text("project_id")
-      .notNull()
-      .references(() => projects.id, { onDelete: "cascade" }),
-    meshPeerId: text("mesh_peer_id").references(() => meshPeers.id, {
-      onDelete: "set null",
-    }),
-    environment: text("environment").notNull(), // production, staging, development, or custom
-    gitRef: text("git_ref"),
-    composeContent: text("compose_content"), // snapshot of compose config at transfer time
-    sourceInstanceId: text("source_instance_id"), // which instance this deployment originated from
-    transferredAt: timestamp("transferred_at"), // when the transfer happened
-    status: appStatusEnum("status").notNull().default("stopped"),
-    lastDeployedAt: timestamp("last_deployed_at"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (t) => [
-    unique("project_instance_peer_env_uniq").on(
-      t.projectId,
-      t.meshPeerId,
-      t.environment
-    ),
-    index("project_instance_project_idx").on(t.projectId),
-    index("project_instance_peer_idx").on(t.meshPeerId),
-  ]
-);
-
-export const projectInstancesRelations = relations(
-  projectInstances,
-  ({ one }) => ({
-    project: one(projects, {
-      fields: [projectInstances.projectId],
-      references: [projects.id],
-    }),
-    meshPeer: one(meshPeers, {
-      fields: [projectInstances.meshPeerId],
-      references: [meshPeers.id],
-    }),
-  })
-);
