+++ lib/db/schema/index.ts
@@ -1,0 +1,13 @@
+export * from "./enums";
+export * from "./auth";
+export * from "./organizations";
+export * from "./projects";
+export * from "./apps";
+export * from "./config";
+export * from "./backups";
+export * from "./notifications";
+export * from "./cron";
+export * from "./mesh";
+export * from "./monitoring";
+export * from "./templates";
+export * from "./relations";
