+++ lib/db/schema/projects.ts
@@ -1,0 +1,28 @@
+import {
+  pgTable,
+  text,
+  timestamp,
+  unique,
+} from "drizzle-orm/pg-core";
+import { organizations } from "./organizations";
+
+// ---------------------------------------------------------------------------
+// Projects (groups of related apps)
+// ---------------------------------------------------------------------------
+
+export const projects = pgTable(
+  "project",
+  {
+    id: text("id").primaryKey(),
+    organizationId: text("organization_id")
+      .notNull()
+      .references(() => organizations.id, { onDelete: "cascade" }),
+    name: text("name").notNull(),
+    displayName: text("display_name").notNull(),
+    description: text("description"),
+    color: text("color").default("#6366f1"),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().notNull(),
+  },
+  (t) => [unique("project_org_name_uniq").on(t.organizationId, t.name)]
+);
